feat(Frontende/Backend): Reservations

+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/service/RevisionService.java
@@ -4,7 +4,7 @@ import com.isa.FishingBooker.model.Revision;
 
 public interface RevisionService extends Service<Revision> {
 	
-	public double getBoatAppointmentRevisionsRate(Integer id);
-	public double getResortAppointmentRevisionsRate(Integer id);
-	public double getTutorServiceAppointmentRevisionsRate(Integer id);
+	public int getBoatAppointmentRevisionsRate(Integer id);
+	public int getResortAppointmentRevisionsRate(Integer id);
+	public int getTutorServiceAppointmentRevisionsRate(Integer id);
 }
